requirements.txt
# FastAPI Framework & ASGI Server
fastapi
uvicorn

# Database & Authentication
pymongo
passlib[bcrypt]
python-dotenv
pyjwt

# AI & NLP
langchain
sentence-transformers
# chromadb
transformers

# PDF Processing
pymupdf

# Role-Based Access Control (RBAC)
oso

# Testing
pytest

# Validation
email-validator
